Make == and eql? more Ruby-like
From:
http://groups.google.com/group/mongomapper/t/85af8026a2dba34d

This is how I would expect an equality test to work in MongoMapper: 
  # Given Document, a MongoMapper Document with a :title key 
  @doc = Document.new 
  @doc_copy = @doc.dup 
  @doc.title = "changed" 
  assert_not_equal @doc_copy, @doc 

My rationale: @doc and @doc_copy should be unequal because not all of 
their key values have the same value. (Also of note: this would match 
the styles in ActiveRecord and DataMapper.) 

To double check my thinking, I did some research. Here is a great 
summary of identity and equality in Ruby: 
http://kentreis.wordpress.com/2007/02/08/identity-and-equality-in-rub... 

Currently, the MongoMapper == implementation looks like this: 
  def ==(other) 
    other.is_a?(self.class) && id == other.id 
  end 
I notice two things wrong with this: 
1. it doesn't check the underlying key values 
2. it does a type check. equality in Ruby using == should not test 
type or class. 

I'm going to work on a patch so that MongoMapper's == checks the 
underlying key values.

One more thing. There is a time and place for checking type, but it 
isn't ==. It is "eql?". So I plan on implementing eql? as well. From 
the link I showed earlier: eql? is "True if the receiver and the 
argument have both the same type and equal values." 

Regarding the built-in '_id' property, I'm going to argue it shouldn't 
be considered when checking equality with ==. 

I consider the MongoMapper 'id' field to be analogous to Ruby's 
#object_id. Ruby's == doesn't care if object_id is different. By 
analogy, MongoMapper's == shouldn't care if '_id' is different. 

To explain the Ruby example in detail: 
  s1 = "abc" 
  s2 = "abc" 
  assert_equal s1, s2 
  assert_not_equal s1.object_id, s2.object_id